+++ x/cronos/keeper/keeper.go
@@ -317,6 +317,9 @@ func (k Keeper) IBCOnAcknowledgementPacketCallback(
 	if err := k.cdc.UnmarshalJSON(acknowledgement, &ack); err != nil {
 		return err
 	}
+	if !ack.Success() {
+		return k.onPacketResult(ctx, packet, false, relayer, contractAddress, packetSenderAddress)
+	}
 	var res channeltypes.Acknowledgement
 	if err := k.cdc.UnmarshalJSON(ack.AppAcknowledgement, &res); err != nil {
 		return err